Get specs passing again for SSL results parsing
This deals with #5319, but does not fix it completely. See the
discussion over there.

Regardless of the partial fix, unless you want to pick up implementing
40-bit scanning and parsing in Rex::SSL directly, I suggest landing this
to get out of the failing spec situation. Some day, Travis-CI will also
fix their system libraries, which will result in builds permafailing
without a reasonable spec change or proper workaround.

Living with known failing tests is living in a state of sin.