Ensure the gpg message is in the right place
Incidentally, also ensure that this is a merge commit we're testing. If
it's just a bare commit, don't bother testing, since that means this
is just a pull request, and not a merge commit.

Now, if the last thing on the pull request /is/ a merge commit (which
can absolutely happen), that must be signed with a good key as well.

So, how can I tell if this is a pull request test or a merge-to-master
test? Is there a way to tell the difference in Travis?