Create enum_application_artefacts_packrat.rb
This post exploitation module has an extensive list of artefacts that it can gather from applications on end user systems.

Artefacts include: chat logins and logs, browser logins and history and cookies, email logins and emails sent and received and deleted, contacts, and many others. These artefacts are collected from applications including: 12 browsers, 13 chat/IM/IRC applications, 6 email clients, and 1 game.

The use case for this post-exploitation module is to specify the types of artefacts you are interested in, to gather the relevant files depending on your aims.