Add additional obfuscation to powershell_vba
This commit adds an additional function in VBA which XORs the input
string with a provided key, which is also provided as a string. The
key used in the XOR process is a randomly generated string created upon
each generation of the payload. The additional pieces which are obfuscated
are powershell.exe along with the arguments passed to it but does not
include the actual payload.

Additionally, rather than use the call to Shell directly I modified the
VBA code to create a WScript.Shell object and then call the Run method on
that and pass the powershell command to run. The WScript.Shell string is
also obfuscated with the XOR function.

All of the added variables as well as function are also generated
randomly upon payload creation.