Add user agent list, modify existing stuff to use it
This commit adds a user_agent module which includes a list of common user agents. The module intends to allow for exploits and modules to refererence user agents easily. The module exposes a full list of agents, a function to select a random agent, a function to select the most common agent, and another function which returns the shortest possible agent (used for Meterpreter payloads).

Existing modules have been updated to reference this new module so that we have a single-stop shop. There might be more cases where this is required, but I was hesitate to change some of the exploits in case those user agents were important.