CatPatch
The AllowPowershellPrompt should be set to true by default. Fixes a bug when it's used in browser_autopwn2 and the "victim" is accessing the exploit with Win7. 
I have noticed that if the AllowPowershellPrompt is set to false, in case of "victim" using win7 the exploit hits send_not_found(cli) on line 363.

Mew.