This is a combination of 5 commits. @db tests not working because the part where we tell it to connect to the database was commented out.
There was a race condition causing the rc script to exit before all of the sessions were opened. This change properly waits for us to get the expected number of sessions before exiting.

Use a more reliable form of the rc script.

Didn't mean to commit the schema.rb

Forgot new line