Use and() and not or() to match the slow search
For some frustrating reason, the db-backed search behaves completely differently than the slow search. Slow search does an `and()`, which makes a lot of sense; the db-backed search does an `or()`, which makes no sense. The end result is that trying to run the search query from the sample help text (`search cve:2009 type:exploit app:client`) returns thousands of results when connected to a DB.

This has supposedly been an issue back to 2013: https://community.rapid7.com/thread/3601