Make vulns and notes with -o similar to services and hosts
Now all of them will simply save the requested information to an output
file if specified and will let you know where the file is when done.
Previously, notes and vulns would also print out all of the notes and
vulns, which just felt weird and inconsistent.  So, I fixed it.