Exclude Ruby external modules from rspec
The sum of the testing code and loading code assumptions was that all
files with the same extension in the same folder were all loadable with
the same loader. This is no longer the case, and until we are ready to
test the load-ability of external modules we can safely ignore them.