Add office365userenum.py as external module
External python module compatible with v2 and v3.

Enumerate valid usernames (email addresses) from Office 365 using
ActiveSync.
Differences in the HTTP Response code and HTTP Headers can be used to
differentiate between:

 - Valid Username (Response code 401)

 - Valid Username and Password without 2FA (Response Code 200)

 - Valid Username and Password with 2FA (Response Code 403)

 - Invalid Username (Response code 404 with Header X-CasErrorCode: UserNotFound)

Note this behaviour appears to be limited to Office365, MS Exchange
does not appear to be affected.

Microsoft Security Response Center stated on 2017-06-28 that this issue
does not "meet the bar for security servicing". As such it is not
expected to be fixed any time soon.

This script is maintaing the ability to run independently of MSF.