Add PTY option to Net::SSH::CommandStream
This allows us to spawn a PTY for our shell session. Note that this will
write us to {u,w}tmp and lastlog, so use this option with care.

And yes, I did change the API, but up until now, we've been using only
the first two parameters. We should be using keyword args. /shrug