This post exploitation module has an extensive list of artefacts that it can gather from applications on end user systems.
Artifacts include: chat logins and logs, browser logins and history and
cookies, email logins and emails sent and received and deleted, contacts
and many others. These artifacts are collected from applications
including: 12 browsers, 13 chat/IM/IRC applications, 6 email clients and
1 game.

These artifacts are then scraped for credentials,

The use case for this post-exploitation module is to specify the types
of artefacts you are interesed in, to gather the relevent files
depending on your aims.