Refactor proxy relay code
The relay module is re-implemented at least three times in Rex
for SOCKS4a, SOCKS5, and HTTP proxies with contextual thread names
as the only differentiator.
Extract the relay code into its own mixin within the proto/proxy
space, include it in the existing consumers, and update to take a
relay thread name as the 3rd argument. Update all #relay method
callers to pass a thread name relevant to the context.