Threaded logger and CONNECT permissions
Implement response logging/website reporting in a separate thread
using duplicates of the original inline req/resp pair and without
referencing the client object which is concurrently being used to
communicate with the remote system. This should help improve
latency and throughput for logged module executions by pulling the
database access out-of-line.

Implement a callback in the Rex Proxy to verify whether or not a
CONNECT from one sockaddr to another is permissible by saddr:sport
daddr:dport. Expose the callback in the Msf namespace, implement
the actual permissions check logic in the MetasploitModule context.

Testing:
  Basic internal tests - CONNECT works as before with defaults,
stops connections returning a 405 when the access control logic
does not find an appropriate entry.