Fallback to TLS MITM when CONNECT is not permitted
HTTP CONNECT proxies are not very useful for attackers as they do
not permit active (MITM) or passive (sniffing) interaction with
the HTTP layer of the (usually) subsequent HTTP/TLS connection.
The verb/semantic provides arbitrary port forwarding capability to
the connecting client and must therefore be restricted in order to
prevent abuse of the attacker's existing pivots/switchboard by
their intended victim.

With CONNECT permissions filtering in place, a graceful fallback
is required (instead of a 405 or 503). Since the vast majority of
HTTP CONNECT requests result in a subsequent TLS negotiation over
the produced TCP port-forward, a simulation of this step by the
proxy itself permits control over the encryption layer and access
to the tunnel's contents.

By reading the Host header from the client's request, a TLS ctx
can be created for the name using rex-socket/pull/17, attached to
the client socket, and set to negotiate a new context with the
actual client. Once this is complete, data is read and written
via the TLS context and HTTP requests are handled as with a plain
text connection.

Testing:
  Curl of https://bbc.com with CONNECT permitted and denied