Update android.rb
the codes above for the dump_whatsapp_enum and the dump_whatsapp features. Here is where the orders are created and the bundles sent to the Meterpreter “server” running through the remote device. In order to understand what’s going on while shipping this orders and the received information, it is highly recommended to check the Packet.rb code.