Fix fail_with check
If the first argument was on a new line, it would fail the check even though the first argument was in the accepted list. This fix concatenates the following line before doing the regex. Also, a `.*` has been changed to `\s` in order to match the new line without having to use the `m` flag. An additional benefit of this is that (while invalid syntax) `fail_with("Failure::None,` will now also be caught.

This will still still cause an error to be reported if there is a blank line between the opening bracket and the first argument, but that seems unlikely to happen in the real world.